Move creating Build Menu into a seperate scene
